feat: bufferevent finalize callback
This callback provides the capacity to free resources. It makes
sure that all bufferevent pending events have finished as well as
all HTTP and SSL resources freed, in case those are used.

This is not possible in the regular callback, as bufferevent still
has pending events by the time its called.